+++ neo/idlib/Str.cpp
@@ -703,7 +703,7 @@ idStr::FileNameHash
 */
 int idStr::FileNameHash( void ) const {
 	int		i;
-	long	hash;
+	int		hash;
 	char	letter;
 
 	hash = 0;
@@ -716,7 +716,7 @@ int idStr::FileNameHash( void ) const {
 		if ( letter =='\\' ) {
 			letter = '/';
 		}
-		hash += (long)(letter)*(i+119);
+		hash += (int)(letter)*(i+119);
 		i++;
 	}
 	hash &= (FILE_HASH_SIZE-1);
